+++ packages/react-cosmos-render/src/CaptureProps.js
@@ -7,7 +7,7 @@ import { extractValuesFromObject } from './shared/values';
 import { getInstanceId, getComponentName } from './shared/decorator';
 
 import type { Element } from 'react';
-import type { FixtureState, SetFixtureState } from './types';
+import type { FixtureState, SetFixtureState } from './types/fixture';
 
 type Props = {
   children: Element<any>
